+++ LaTeX.novaextension/Queries/bibtex/textChecking.scm
@@ -1,0 +1,5 @@
+[
+  (brace_word)
+  (quote_word)
+  (junk)
+] @subtree