+++ LaTeX.novaextension/Queries/latex/tagMatching.scm
@@ -1,0 +1,8 @@
+(_
+  (begin
+    name: (curly_group_text
+      (text) @tag.start.name)) @tag.start
+  (end
+    name: (curly_group_text
+      (text) @tag.end.name)) @tag.end
+)