+++ LaTeX.novaextension/Queries/latex/textChecking.scm
@@ -1,0 +1,6 @@
+[
+  (word)
+  (line_comment)
+  (block_comment)
+  (comment_environment)
+] @subtree